Change the order of initialization for the Application class' members
Certain Component.onDestruction handlers may require access to the Scene
instance while the UserInterface object is being torn down and so it is
trivial to see why having the scene destroyed before the UserInterface
object is problematic.

To fix this, the userInterface_ member variable is instantiated last, and
therefore destroyed first.