Modify the FramebufferProvider::requestImage's behavior
Prior to the introduction of the FramebufferView component, an Image
component was used to view the content of the framebuffer. However,
Image components do not account for frequent updates to an image's pixels
so a workaround was introduced in @5d12f4e1fa to mitigate this problem.

While functional, the workaround was still slow and frame transitions
were not fluid, i.e. there was a slight "flash" between each frame.
So the FramebufferView component was introduced in @51cf4464d9 to solve
this problem.

Since the Image component is no longer used to view the framebuffer, the
workaround has been removed from the requestImage function.