Improve the renderer's rasterization step
By definition, the rasterization process should return a collection of
fragments and so the old implementation did just that: return a list of
fragments computed from a set of primitives. As you can imagine, hundreds
(thousands even) of fragments can be generated for a single polygon face
and the overhead of creating a new list is not negligible.

The new implementation no longer creates a list of fragments but
immediately processes a fragment the instant it is created.