Change the FramebufferView's texture filters
Using the nearest-neighbor filter (GL_NEAREST) provides a faster render
however when the view has a small resolution, some texels are discared.

For instance, this is very apparent when the renderer's output is
minimized while the rendering context is being configured; or on screens
with small resolutions (e.g. 1024x768).

This problem is fixed by simply using (bi)linear filters.